+++ backend/test/AgentNFT.t.sol
@@ -24,6 +24,7 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
         uint128 creationTime;
         uint128 lastUpdate;
     }
+
     mapping(uint256 => AgentMetadata) private _agentMetadata;
 
     // Access control: Authorized minters (for breeding contract, etc.)
@@ -34,22 +35,10 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     uint256 public constant MAX_MEMORY_LOGS_PER_AGENT = 1000;
 
     // Events
-    event AgentMinted(
-        uint256 indexed tokenId,
-        address indexed owner,
-        string tokenUri
-    );
+    event AgentMinted(uint256 indexed tokenId, address indexed owner, string tokenUri);
     event MemoryHashSet(uint256 indexed tokenId, string memoryHash);
-    event MemoryLogged(
-        uint256 indexed tokenId,
-        bytes32 indexed hash,
-        uint256 timestamp
-    );
-    event MemoryForgotten(
-        uint256 indexed tokenId,
-        bytes32 indexed hash,
-        uint256 timestamp
-    );
+    event MemoryLogged(uint256 indexed tokenId, bytes32 indexed hash, uint256 timestamp);
+    event MemoryForgotten(uint256 indexed tokenId, bytes32 indexed hash, uint256 timestamp);
     event TraitsUpdated(uint256 indexed tokenId, string[] traits);
     event AvatarUpdated(uint256 indexed tokenId, string avatarHash);
     event AuthorizedMinterAdded(address indexed minter);
@@ -58,28 +47,20 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     constructor() ERC721("AgentNFT", "AGNT") Ownable(msg.sender) {}
 
     modifier onlyOwnerOrAuthorized() {
-        require(
-            msg.sender == owner() || authorizedMinters[msg.sender],
-            "AgentNFT: Not authorized"
-        );
+        require(msg.sender == owner() || authorizedMinters[msg.sender], "AgentNFT: Not authorized");
         _;
     }
 
     modifier validTokenId(uint256 tokenId) {
-        require(
-            _ownerOf(tokenId) != address(0),
-            "AgentNFT: Token does not exist"
-        );
+        require(_ownerOf(tokenId) != address(0), "AgentNFT: Token does not exist");
         _;
     }
 
     modifier onlyTokenOwnerOrApproved(uint256 tokenId) {
         address tokenOwner = ownerOf(tokenId);
         require(
-            msg.sender == tokenOwner ||
-                msg.sender == owner() ||
-                getApproved(tokenId) == msg.sender ||
-                isApprovedForAll(tokenOwner, msg.sender),
+            msg.sender == tokenOwner || msg.sender == owner() || getApproved(tokenId) == msg.sender
+                || isApprovedForAll(tokenOwner, msg.sender),
             "AgentNFT: Not authorized"
         );
         _;
@@ -98,10 +79,7 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Mint a new Agent NFT - only owner or authorized minters
-    function mint(
-        address to,
-        string memory tokenUri
-    ) external onlyOwnerOrAuthorized nonReentrant {
+    function mint(address to, string memory tokenUri) external onlyOwnerOrAuthorized nonReentrant {
         require(to != address(0), "AgentNFT: Mint to zero address");
         require(bytes(tokenUri).length > 0, "AgentNFT: Empty token URI");
 
@@ -110,25 +88,17 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
         _setTokenURI(tokenId, tokenUri);
 
         // Set metadata
-        _agentMetadata[tokenId] = AgentMetadata({
-            creationTime: uint128(block.timestamp),
-            lastUpdate: uint128(block.timestamp)
-        });
+        _agentMetadata[tokenId] =
+            AgentMetadata({creationTime: uint128(block.timestamp), lastUpdate: uint128(block.timestamp)});
 
         emit AgentMinted(tokenId, to, tokenUri);
         nextTokenId++;
     }
 
     // Public mint function with traits initialization and payment (optional)
-    function publicMint(
-        string memory tokenUri,
-        string[] memory initialTraits
-    ) external payable nonReentrant {
+    function publicMint(string memory tokenUri, string[] memory initialTraits) external payable nonReentrant {
         require(bytes(tokenUri).length > 0, "AgentNFT: Empty token URI");
-        require(
-            initialTraits.length <= MAX_TRAITS_PER_AGENT,
-            "AgentNFT: Too many traits"
-        );
+        require(initialTraits.length <= MAX_TRAITS_PER_AGENT, "AgentNFT: Too many traits");
 
         uint256 tokenId = nextTokenId;
         _safeMint(msg.sender, tokenId);
@@ -138,10 +108,8 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
         _agentTraits[tokenId] = initialTraits;
 
         // Set metadata
-        _agentMetadata[tokenId] = AgentMetadata({
-            creationTime: uint128(block.timestamp),
-            lastUpdate: uint128(block.timestamp)
-        });
+        _agentMetadata[tokenId] =
+            AgentMetadata({creationTime: uint128(block.timestamp), lastUpdate: uint128(block.timestamp)});
 
         emit AgentMinted(tokenId, msg.sender, tokenUri);
         emit TraitsUpdated(tokenId, initialTraits);
@@ -149,10 +117,7 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Set or update memory hash for a given tokenId with reentrancy protection
-    function setMemoryHash(
-        uint256 tokenId,
-        string memory hash
-    )
+    function setMemoryHash(uint256 tokenId, string memory hash)
         external
         validTokenId(tokenId)
         onlyTokenOwnerOrApproved(tokenId)
@@ -167,22 +132,16 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Set agent traits with validation
-    function setTraits(
-        uint256 tokenId,
-        string[] memory traits
-    )
+    function setTraits(uint256 tokenId, string[] memory traits)
         external
         validTokenId(tokenId)
         onlyTokenOwnerOrApproved(tokenId)
         nonReentrant
     {
-        require(
-            traits.length <= MAX_TRAITS_PER_AGENT,
-            "AgentNFT: Too many traits"
-        );
+        require(traits.length <= MAX_TRAITS_PER_AGENT, "AgentNFT: Too many traits");
 
         // Validate trait strings are not empty
-        for (uint i = 0; i < traits.length; i++) {
+        for (uint256 i = 0; i < traits.length; i++) {
             require(bytes(traits[i]).length > 0, "AgentNFT: Empty trait");
         }
 
@@ -193,10 +152,7 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Set avatar hash with validation
-    function setAvatarHash(
-        uint256 tokenId,
-        string memory avatarHash
-    )
+    function setAvatarHash(uint256 tokenId, string memory avatarHash)
         external
         validTokenId(tokenId)
         onlyTokenOwnerOrApproved(tokenId)
@@ -211,30 +167,21 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Log memory hash for audit trail with limits
-    function logMemoryHash(
-        uint256 tokenId,
-        bytes32 hash
-    )
+    function logMemoryHash(uint256 tokenId, bytes32 hash)
         external
         validTokenId(tokenId)
         onlyTokenOwnerOrApproved(tokenId)
         nonReentrant
     {
         require(hash != bytes32(0), "AgentNFT: Invalid hash");
-        require(
-            _memoryAuditLog[tokenId].length < MAX_MEMORY_LOGS_PER_AGENT,
-            "AgentNFT: Memory log limit reached"
-        );
+        require(_memoryAuditLog[tokenId].length < MAX_MEMORY_LOGS_PER_AGENT, "AgentNFT: Memory log limit reached");
 
         _memoryAuditLog[tokenId].push(hash);
         emit MemoryLogged(tokenId, hash, block.timestamp);
     }
 
     // Log memory deletion for audit trail
-    function logMemoryForgotten(
-        uint256 tokenId,
-        bytes32 hash
-    )
+    function logMemoryForgotten(uint256 tokenId, bytes32 hash)
         external
         validTokenId(tokenId)
         onlyTokenOwnerOrApproved(tokenId)
@@ -245,39 +192,27 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // View functions
-    function getMemoryHash(
-        uint256 tokenId
-    ) external view validTokenId(tokenId) returns (string memory) {
+    function getMemoryHash(uint256 tokenId) external view validTokenId(tokenId) returns (string memory) {
         return _memoryHashes[tokenId];
     }
 
-    function getTraits(
-        uint256 tokenId
-    ) external view validTokenId(tokenId) returns (string[] memory) {
+    function getTraits(uint256 tokenId) external view validTokenId(tokenId) returns (string[] memory) {
         return _agentTraits[tokenId];
     }
 
-    function getAvatarHash(
-        uint256 tokenId
-    ) external view validTokenId(tokenId) returns (string memory) {
+    function getAvatarHash(uint256 tokenId) external view validTokenId(tokenId) returns (string memory) {
         return _agentAvatarHash[tokenId];
     }
 
-    function getMemoryAuditLog(
-        uint256 tokenId
-    ) external view validTokenId(tokenId) returns (bytes32[] memory) {
+    function getMemoryAuditLog(uint256 tokenId) external view validTokenId(tokenId) returns (bytes32[] memory) {
         return _memoryAuditLog[tokenId];
     }
 
-    function getMemoryAuditLogCount(
-        uint256 tokenId
-    ) external view validTokenId(tokenId) returns (uint256) {
+    function getMemoryAuditLogCount(uint256 tokenId) external view validTokenId(tokenId) returns (uint256) {
         return _memoryAuditLog[tokenId].length;
     }
 
-    function getAgentMetadata(
-        uint256 tokenId
-    )
+    function getAgentMetadata(uint256 tokenId)
         external
         view
         validTokenId(tokenId)
@@ -292,11 +227,9 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Gas optimization: Batch operations
-    function getMultipleTraits(
-        uint256[] calldata tokenIds
-    ) external view returns (string[][] memory) {
+    function getMultipleTraits(uint256[] calldata tokenIds) external view returns (string[][] memory) {
         string[][] memory results = new string[][](tokenIds.length);
-        for (uint i = 0; i < tokenIds.length; i++) {
+        for (uint256 i = 0; i < tokenIds.length; i++) {
             if (_ownerOf(tokenIds[i]) != address(0)) {
                 results[i] = _agentTraits[tokenIds[i]];
             }
@@ -315,7 +248,7 @@ contract AgentNFT is ERC721URIStorage, Ownable, ReentrancyGuard {
     }
 
     // Upgrade safety: Prevent accidental burning
-    function burn(uint256 /*tokenId*/) external pure {
+    function burn(uint256 /*tokenId*/ ) external pure {
         require(false, "AgentNFT: Burning not allowed");
     }
 }
